+++ app/controllers/actors_controller.rb
@@ -2,6 +2,8 @@ class ActorsController < ApplicationController
     before_action :authenticate_user!, only: [:admin, :destroy, :create, :edit, :update]
 
     def admin
+        return unless is_admin?
+
         actors = Actor.select(:id, :name, :surname).order(created_at: :desc).offset((params[:page].to_i-1)*10).limit(11)
 
         render json: {
@@ -12,6 +14,8 @@ class ActorsController < ApplicationController
     end
 
     def destroy
+        return unless is_admin?
+
         actor = Actor.find(params[:id])
         status = 400
         status = 200 if actor.present? && actor.destroy
@@ -20,6 +24,8 @@ class ActorsController < ApplicationController
     end
 
     def create
+        return unless is_admin?
+
         actor = Actor.new(params.permit!.except(:controller, :action))
         status = actor.save ? 200 : 400
 
@@ -28,6 +34,8 @@ class ActorsController < ApplicationController
     end
 
     def edit
+        return unless is_admin?
+
         actor = Actor.find(params[:id])
 
         render json: {
@@ -38,6 +46,8 @@ class ActorsController < ApplicationController
     end
 
     def update
+        return unless is_admin?
+        
         actor = Actor.find(params[:id])
         if actor.present?
             actor.name = params[:name] if params.key?(:name)
