+++ app/controllers/showtimes_controller.rb
@@ -2,6 +2,8 @@ class ShowtimesController < ApplicationController
     before_action :authenticate_user!, only: [:admin, :destroy, :create, :edit, :update]
 
     def admin
+        return unless is_admin?
+
         sql_query = "SELECT s.id, m.name AS movie_name, t.name AS theatre_name, s.projection_date, TIME_FORMAT(s.projection_time, '%H:%i') AS projection_time
             FROM showtimes s
                 JOIN movies m ON s.movie_id = m.id
@@ -21,6 +23,8 @@ class ShowtimesController < ApplicationController
     end
 
     def destroy
+        return unless is_admin?
+
         showtime = Showtime.find(params[:id])
         status = 400
         status = 200 if showtime.present? && showtime.destroy
@@ -29,6 +33,8 @@ class ShowtimesController < ApplicationController
     end
 
     def create
+        return unless is_admin?
+
         if Movie.find(params[:movie_id]).blank? && Theatre.find(params[:theatre_id]).blank?
             status = 400
         else
@@ -40,6 +46,8 @@ class ShowtimesController < ApplicationController
     end
 
     def edit
+        return unless is_admin?
+
         showtime = Showtime.find(params[:id])
         showtime.projection_time = showtime.projection_time.strftime("%H:%M")
 
@@ -51,6 +59,8 @@ class ShowtimesController < ApplicationController
     end
 
     def update
+        return unless is_admin?
+        
         showtime = Showtime.find(params[:id])
         if showtime.present? && Movie.find(params[:movie_id]).present? && Theatre.find(params[:theatre_id]).present?
             showtime.movie_id = params[:movie_id] if params.key?(:movie_id)
